Renamed the tools module.  Fixed the tests.
